@@ -19,7 +19,7 @@ include = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.6.2"
+python = "^3.7"
 "datanommer.models" = "^0.9.1"
 fedmsg = "^1.1.2"
 
@@ -33,6 +33,9 @@ pytest = "^6.2.4"
 fedmsg-meta-fedora-infrastructure = "^0.31.0"
 liccheck = "^0.6.0"
 pytest-cov = "^2.12.1"
+pytest-mock = "^3.6.1"
+pytest-postgresql = "^3.1.1"
+psycopg2 = "^2.9.1"
 
 [tool.poetry.scripts]
 datanommer-create-db = "datanommer.commands:create"
