@@ -19,10 +19,11 @@ include = [
 ]
 
 [tool.poetry.dependencies]
-python = "^3.6.2"
+python = "^3.7"
 fedmsg = "^1.1.2"
 SQLAlchemy = ">=1.0.9"
 alembic = "^1.6.5"
+psycopg2 = "^2.9.1"
 
 [tool.poetry.dev-dependencies]
 pre-commit = "^2.13.0"
@@ -33,7 +34,8 @@ pytest = "^6.2.4"
 fedmsg-meta-fedora-infrastructure = "^0.31.0"
 liccheck = "^0.6.0"
 pytest-cov = "^2.12.1"
-
+pytest-postgresql = "^3.1.1"
+pytest-mock = "^3.6.1"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
