@@ -15,5 +15,5 @@
 # with this program.  If not, see <http://www.gnu.org/licenses/>.
 config = {
     "datanommer.enabled": True,
-    "datanommer.sqlalchemy.url": "sqlite:///datanommer.db",
+    "datanommer.sqlalchemy.url": "postgresql://datanommer:datanommer@localhost/messages",
 }
